+++ src/steps/active-directory/__tests__/converters.test.ts
@@ -192,8 +192,7 @@ describe('createGroupMemberRelationship', () => {
       createGroupMemberRelationship(exampleGroupEntity, exampleGroupMemberUser),
     ).toEqual({
       _class: 'HAS',
-      _key:
-        '89fac263-2430-48fd-9278-dacfdfc89792|has|324e8daa-9c29-42a4-a74b-b9893e6d9750',
+      _key: '89fac263-2430-48fd-9278-dacfdfc89792|has|324e8daa-9c29-42a4-a74b-b9893e6d9750',
       _type: 'azure_group_has_member',
       _mapping: {
         relationshipDirection: RelationshipDirection.FORWARD,
@@ -202,7 +201,7 @@ describe('createGroupMemberRelationship', () => {
         targetEntity: {
           _key: '324e8daa-9c29-42a4-a74b-b9893e6d9750',
           _type: 'azure_user',
-          _class: 'User',
+          _class: ['User'],
           displayName: 'User Name',
           jobTitle: 'Job Title',
           email: 'user@example.com',
@@ -223,8 +222,7 @@ describe('createGroupMemberRelationship', () => {
       ),
     ).toEqual({
       _class: 'HAS',
-      _key:
-        '89fac263-2430-48fd-9278-dacfdfc89792|has|324e8daa-9c29-42a4-a74b-b9893e6d9750',
+      _key: '89fac263-2430-48fd-9278-dacfdfc89792|has|324e8daa-9c29-42a4-a74b-b9893e6d9750',
       _type: 'azure_group_has_member',
       _mapping: {
         relationshipDirection: RelationshipDirection.FORWARD,
@@ -233,7 +231,7 @@ describe('createGroupMemberRelationship', () => {
         targetEntity: {
           _key: '324e8daa-9c29-42a4-a74b-b9893e6d9750',
           _type: 'azure_user_group',
-          _class: 'UserGroup',
+          _class: ['UserGroup'],
           displayName: 'Managers',
           jobTitle: null,
           email: null,
@@ -254,8 +252,7 @@ describe('createGroupMemberRelationship', () => {
       ),
     ).toEqual({
       _class: 'HAS',
-      _key:
-        '89fac263-2430-48fd-9278-dacfdfc89792|has|324e8daa-9c29-42a4-a74b-b9893e6d9750',
+      _key: '89fac263-2430-48fd-9278-dacfdfc89792|has|324e8daa-9c29-42a4-a74b-b9893e6d9750',
       _type: 'azure_group_has_member',
       _mapping: {
         relationshipDirection: RelationshipDirection.FORWARD,