+++ src/steps/intune/steps/applications/index.ts
@@ -4,6 +4,7 @@ import {
   createDirectRelationship,
   JobState,
   Entity,
+  IntegrationLogger,
 } from '@jupiterone/integration-sdk-core';
 import { IntegrationConfig, IntegrationStepContext } from '../../../../types';
 import {
@@ -13,11 +14,11 @@ import {
   steps,
 } from '../../constants';
 import {
+  buildDetectedApplicationEntityKey,
   createDetectedApplicationEntity,
+  createDeviceInstalledApplicationRelationship,
   createManagedApplicationEntity,
-  DETECTED_APP_KEY_PREFIX,
   findNewestVersion,
-  UNVERSIONED,
 } from './converters';
 import { DeviceManagementIntuneClient } from '../../clients/deviceManagementIntuneClient';
 import { DetectedApp } from '@microsoft/microsoft-graph-types-beta';
@@ -89,6 +90,18 @@ export async function fetchManagedApplications(
   });
 }
 
+type DebugParams = {
+  logger: IntegrationLogger;
+  args: Record<string, any>;
+  msg: string;
+};
+
+function debug({ logger, args, msg }: DebugParams) {
+  if (process.env.DEBUG_APPLICATIONS) {
+    logger.info(args, msg);
+  }
+}
+
 /**
  * Creates a single `Application { _type: 'intune_detected_application' }` entity generated
  * for each `Application.displayName`. All `Device` entities which have an app installed with
@@ -103,34 +116,59 @@ export async function fetchDetectedApplications(
     logger,
     instance.config,
   );
+
+  const deviceEntityIdSet = new Set<string>();
+
   for (const type of managedDeviceTypes) {
     await jobState.iterateEntities({ _type: type }, async (deviceEntity) => {
+      const deviceEntityId = deviceEntity.id as string;
+
+      if (deviceEntityIdSet.has(deviceEntityId)) {
+        debug({
+          logger,
+          args: {
+            deviceEntityId,
+          },
+          msg: 'Found duplicate device entity ID',
+        });
+      } else {
+        deviceEntityIdSet.add(deviceEntityId);
+      }
+
       await intuneClient.iterateDetectedApps(
-        deviceEntity.id as string,
+        deviceEntityId,
         async ({ detectedApps }) => {
           for (const detectedApp of detectedApps ?? []) {
             // Ingest all assigned or line of business apps reguardless if a device has installed it or not yet
-            const detectedAppEntity = await findOrCreateDetectedApplicationEntity(
-              detectedApp,
-              jobState,
-            );
+            const detectedAppEntity =
+              await findOrCreateDetectedApplicationEntity(
+                detectedApp,
+                jobState,
+              );
 
-            const version = detectedApp.version ?? UNVERSIONED;
-            const directRelationship = createDirectRelationship({
-              _class:
-                relationships.MULTI_DEVICE_INSTALLED_DETECTED_APPLICATION[0]
-                  ._class,
-              from: deviceEntity,
-              to: detectedAppEntity,
-              properties: {
-                version,
-                detectionId: detectedApp.id, // unique id for the specific detection
-              },
-            });
-            // Need to append the detectionId to the end of the key so there can be multiple relationships to the same Application entities
-            directRelationship._key += `|${detectedApp.id}`;
-            await jobState.addRelationship(directRelationship);
+            const deviceInstalledRelationship =
+              createDeviceInstalledApplicationRelationship({
+                deviceEntity,
+                detectedAppEntity,
+                detectedApp,
+              });
 
+            if (await jobState.hasKey(deviceInstalledRelationship._key)) {
+              logger.warn(
+                {
+                  relationshipKey: deviceInstalledRelationship._key,
+                  deviceKey: deviceEntity._key,
+                  detectedAppEntityKey: detectedAppEntity._key,
+                  detectedAppId: detectedApp.id,
+                  relationshipClass:
+                    relationships.MULTI_DEVICE_INSTALLED_DETECTED_APPLICATION[0]
+                      ._class,
+                },
+                'Possible duplicate deviceInstalledDetectedApp Key',
+              );
+            } else {
+              await jobState.addRelationship(deviceInstalledRelationship);
+            }
             // TODO create managed -> detected relationships
             // // If there is a managed application related to this, create a MANAGES relationship
             // let managedAppEntity;
@@ -172,16 +210,16 @@ async function findOrCreateDetectedApplicationEntity(
   detectedApp: DetectedApp,
   jobState: JobState,
 ): Promise<Entity> {
-  let detectedAppEntity =
-    detectedApp.id &&
-    (await jobState.findEntity(
-      DETECTED_APP_KEY_PREFIX + detectedApp.displayName?.toLowerCase(),
-    ));
+  let detectedAppEntity = await jobState.findEntity(
+    buildDetectedApplicationEntityKey(detectedApp),
+  );
 
   if (!detectedAppEntity) {
-    detectedAppEntity = createDetectedApplicationEntity(detectedApp);
-    await jobState.addEntity(detectedAppEntity);
+    detectedAppEntity = await jobState.addEntity(
+      createDetectedApplicationEntity(detectedApp),
+    );
   }
+
   return detectedAppEntity;
 }