+++ src/steps/intune/steps/compliancePoliciesAndFindings/converters.ts
@@ -29,8 +29,9 @@ export function createCompliancePolicyEntity(
 ): Entity {
   return createIntegrationEntity({
     entityData: {
-      source: compliancePolicy,
+      source: {},
       assign: {
+        _key: compliancePolicy.id!,
         _class: entities.COMPLIANCE_POLICY._class,
         _type: entities.COMPLIANCE_POLICY._type,
         function: 'endpoint-compliance', // needed to match up with existing integrations data structures.
@@ -58,6 +59,7 @@ export function createNoncomplianceFindingEntity(
     entityData: {
       source: deviceStatus,
       assign: {
+        _key: deviceStatus.id!,
         _class: entities.NONCOMPLIANCE_FINDING._class,
         _type: entities.NONCOMPLIANCE_FINDING._type,
         id: deviceStatus.id,